requirements.txt
spacy==3.7.2
pandas==2.1.4
python-magic==0.4.27
PyPDF2==3.0.1
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
transformers==4.36.2
torch==2.1.2
pydantic==2.5.3 